Handle changing the duty cycle in an ISR
The Pi Pico allows for a IRQ to be raised when the PWM counter rolls
over. That's excellent, it gives me a clean way to update the config on
the PWM slices exactly once per cycle. (The PWM counter only updates the
duty cycle on each wrap of the counter, there's no point in trying to
update it more often.)

This decouples the PWM updates (which means the servo updates) from the
input. This code is for a device that allows moving servos over DMX,
which runs at 40Hz at the most, and can be far less. Having the input
and the action on the servos decoupled should give me a lot of freedom
from here on out, and allow me to make sure the servos are rock solid.