Watch the UART for messages to move the servos
It's speaking a "DMX-like" protocol where it's looking at a packed array
of bytes, and checking that each packet matches to a known header, and
that it ends in 0x00. If it does not end in 0x00 it starts discarding
packets until it finds one, as a form of error handling.

It seems to work pretty well.

Sometimes the Pi Pico is freezeing. I still need to figure out why.