Motion smoothing!!! 🎉
This commit adds software motion smoothing to the servo controller. It
does this by using the configured smoothing value provided in the
config. The higher the value, the smoother it will be.

There's a new task present now. The controller runs a "housekeeping"
task that's fired off after the ISR for the servo. Right now all it's
doing is setting up the next value for the motion smoothing (so that the
math doesn't happen within the ISR), but could be used for lots of
things.