Start the process of having the parrot process input values
In reality we don't want to just mirror what comes in from the input
handler, we want to process it to make it do what _this creature_ should
do with it. This is the start of that.

I've largely written this code before in golang. I'm re-working it here
in C++ for the new controller.