Move the debug shell to the CDC serial port
I need the extra pins for other things. The use of two of them for
a debug shell just isn't great. It's cool being on USB, anyhow, since
I don't need to carry a UART -> USB thingy with me.

I'm having problems overruning the TX buffer, so for now there's a delay
of 1ms per character. It helps me re-live the BBS days!