Tighten up the tolerances on the latch
I went exploring with my logic analyzer. The datasheet says that it
takes 220ns in the worse case (at 2v, when really hot) for the latch to
go. We're running at 3.3v, and not really hot, but I wanted to make sure
I stay within those tolerances.

Having the CPU count to 3 takes about 380ns. This puts the latch delay
well within tolerances, but is still a very tiny delay. I measured it
usually taking about 42ns for the latch to flip, so I'm really
comfortable with using 380ns.

Fun discovery! It counts to 3 faster with a 32-bit int than an 8-bit
one! I guess that makes sense, the Pi Pico is a 32-bit MCU.